# docs/swagger.yaml
basePath: /
definitions:
  game.Answer:
    properties:
      answerIndex:
        description: 0..3
        type: integer
      playerID:
        type: string
      submittedAt:
        type: string
    type: object
  game.Player:
    properties:
      connected:
        type: boolean
      id:
        description: Session ID / Socket ID
        type: string
      nickname:
        type: string
      score:
        type: integer
    type: object
  game.Room:
    properties:
      answers:
        additionalProperties:
          $ref: '#/definitions/game.Answer'
        description: Map[PlayerID]*Answer (da pergunta atual)
        type: object
      currentQuestionIndex:
        type: integer
      id:
        type: string
      players:
        additionalProperties:
          $ref: '#/definitions/game.Player'
        description: Map[PlayerID]*Player
        type: object
      quiz:
        $ref: '#/definitions/quiz.Quiz'
      status:
        type: string
      teacherID:
        type: string
    type: object
...
  quiz.Question:
    properties:
      correctIndex:
        description: 0..3
        type: integer
      createdAt:
        type: string
      id:
        type: string
      optionA:
        description: "0"
        type: string
      optionB:
        description: "1"
        type: string
      optionC:
        description: "2"
        type: string
      optionD:
        description: "3"
        type: string
      prompt:
        description: Enunciado
        type: string
      quizId:
        type: string
      sortOrder:
        description: Ordem na lista
        type: integer
      updatedAt:
        type: string
    type: object
  quiz.Quiz:
    properties:
      createdAt:
        type: string
      description:
        type: string
      grade:
        description: 'Série (ex: 7º Ano)'
        type: string
      id:
        type: string
      questions:
        items:
          $ref: '#/definitions/quiz.Question'
        type: array
      status:
        description: DRAFT | PUBLISHED
        type: string
      subject:
        description: 'Disciplina (ex: História)'
        type: string
      teacherId:
        description: Owner
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  teacher.Teacher:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  usecases.AddQuestionInput:
    properties:
      correctIndex:
        type: integer
      optionA:
        type: string
      optionB:
        type: string
      optionC:
        type: string
      optionD:
        type: string
      prompt:
        type: string
      quizID:
        type: string
      teacherID:
        type: string
    type: object
...
  usecases.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  usecases.LoginOutput:
    properties:
      accessToken:
        type: string
      expiresIn:
        description: Segundos
        format: int64
        type: integer
    type: object
  usecases.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  usecases.RegisterOutput:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  usecases.UpdateQuestionInput:
    properties:
      correctIndex:
        type: integer
      optionA:
        type: string
      optionB:
        type: string
      optionC:
        type: string
      optionD:
        type: string
      prompt:
        type: string
      questionID:
        type: string
      quizID:
        type: string
      teacherID:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Suporte RankIt
  description: Backend para a plataforma RankIt (Gamificação em sala de aula).
  title: RankIt API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Realiza login com email e senha e retorna um token JWT.
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecases.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.LoginOutput'
        "401":
          description: Credenciais inválidas
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Autentica um professor
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Obtém detalhes do perfil do usuário autenticado via token JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teacher.Teacher'
        "401":
          description: Não autenticado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Retorna dados do professor logado
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cria uma conta de professor com nome, email e senha.
      parameters:
      - description: Dados de cadastro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecases.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/usecases.RegisterOutput'
        "400":
          description: Erro de validação
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email já cadastrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cadastra um novo professor
      tags:
      - Auth
...
  /quizzes/{id}/questions:
    post:
      consumes:
      - application/json
      description: Adiciona uma nova pergunta ao quiz em DRAFT.
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      - description: Dados da Pergunta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecases.AddQuestionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/quiz.Question'
        "400":
          description: Dados inválidos
      security:
      - BearerAuth: []
      summary: Adiciona pergunta ao quiz
      tags:
      - Questions
  /quizzes/{id}/questions/{questionId}:
    delete:
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      - description: ID da Pergunta
        in: path
        name: questionId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Remove pergunta
      tags:
      - Questions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      - description: ID da Pergunta
        in: path
        name: questionId
        required: true
        type: string
      - description: Dados da Pergunta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usecases.UpdateQuestionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quiz.Question'
      security:
      - BearerAuth: []
      summary: Atualiza pergunta
      tags:
      - Questions
  /quizzes/{id}/questions/reorder:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID do Quiz
        in: path
        name: id
        required: true
        type: string
      - description: Lista com order (Ids)
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            items:
              type: string
            type: array
          type: object
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Reordena perguntas
      tags:
      - Questions
...
  /rooms:
    post:
      consumes:
      - application/json
      description: Cria uma nova sala a partir de um quiz PUBLISHED.
      parameters:
      - description: 'payload: {quizId: uuid}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/game.Room'
        "400":
          description: Quiz inválido
      security:
      - BearerAuth: []
      summary: Cria uma sala de jogo
      tags:
      - Rooms
  /rooms/{id}:
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/game.Room'
        "404":
          description: Sala não encontrada
      summary: Obtém dados da sala
      tags:
      - Rooms
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"